Extract web localDb database migrations
