+++ apps/web/src/localDb/core/database.ts
@@ -2,20 +2,20 @@ import type {
   Card,
   CloudSettings,
   Deck,
-  LegacyEffortLevel,
   MediaAsset,
   ReviewEvent,
-  SyncPushOperation,
   WorkspaceSchedulerSettings,
 } from "../../types";
-import { appendLegacyEffortTag } from "../../legacyEffort";
-import { deriveDueAtBucketMillis, deriveDueAtMillis, parseDueAtMillis } from "../../appData/domain/dueAt";
 import {
   addWebBreadcrumb,
   type IndexedDbOperation,
   type WebObservationScope,
 } from "../../observability/webObservability";
-import { normalizeCardMetadata, normalizeCardType } from "../../appData/domain/cardMetadata";
+import { upgradeDatabase } from "./databaseMigrations";
+import { databaseName, databaseVersion } from "./databaseSchema";
+import type { DatabaseStores } from "./databaseSchema";
+
+export type { DatabaseStores } from "./databaseSchema";
 
 export type StoredCard = Readonly<{
   workspaceId: string;
@@ -85,18 +85,6 @@ export type ProgressCacheStateRecord = Readonly<{
   updatedAt: string;
 }>;
 
-export type DatabaseStores =
-  | "cards"
-  | "cardTags"
-  | "decks"
-  | "mediaAssets"
-  | "progressDailyCounts"
-  | "reviewEvents"
-  | "workspaceSettings"
-  | "workspaceSyncState"
-  | "outbox"
-  | "meta";
-
 export type IndexedDbOpenLifecycleSnapshot = Readonly<{
   observedAt: string;
   databaseName: string;
@@ -107,69 +95,12 @@ export type IndexedDbOpenLifecycleSnapshot = Readonly<{
   databaseUpgraded: boolean;
 }>;
 
-const databaseName = "flashcards-web-sync";
-const databaseVersion = 17;
-const progressCacheStateKey = "progress_cache_state";
 const deleteDatabaseBlockedWaitMs = 3000;
 const activeDatabaseOperationPromises = new Set<Promise<unknown>>();
 let isDatabaseDeleteInProgress = false;
 let activeDatabaseDeletePromise: Promise<void> | null = null;
 let lastIndexedDbOpenLifecycleSnapshot: IndexedDbOpenLifecycleSnapshot | null = null;
 let lastIndexedDbVersionChangeLifecycleSnapshot: IndexedDbOpenLifecycleSnapshot | null = null;
-
-type StoredCardMetadataMigrationFields = Readonly<{
-  cardType?: unknown;
-  metadata?: unknown;
-}>;
-
-type StoredCardMetadataMigrationRecord = Omit<StoredCard, "cardType" | "metadata"> & StoredCardMetadataMigrationFields;
-
-type StoredCardDueAtMigrationRecord = Omit<
-  StoredCard,
-  "dueAt" | "dueAtMillis" | "dueAtBucketMillis" | "fsrsLastReviewedAtMillis" | "cardType" | "metadata"
-> & StoredCardMetadataMigrationFields & Readonly<{
-  dueAt?: string | null;
-  dueAtMillis?: number | null;
-  dueAtBucketMillis?: number;
-  fsrsLastReviewedAtMillis?: number | null;
-}>;
-
-type StoredCardLegacyEffortMigrationRecord = StoredCardDueAtMigrationRecord & Readonly<{
-  effortLevel?: LegacyEffortLevel;
-}>;
-
-type StoredDeckLegacyEffortMigrationRecord = Omit<Deck, "filterDefinition"> & Readonly<{
-  filterDefinition: Readonly<{
-    version: 2;
-    effortLevels?: ReadonlyArray<LegacyEffortLevel>;
-    tags: ReadonlyArray<string>;
-  }>;
-}>;
-
-type CardUpsertOperation = Extract<
-  SyncPushOperation,
-  Readonly<{ entityType: "card"; action: "upsert" }>
->;
-
-type CardUpsertLegacyEffortMigrationOperation = Omit<CardUpsertOperation, "payload"> & Readonly<{
-  payload: Omit<CardUpsertOperation["payload"], "cardType" | "metadata"> & Readonly<{
-    cardType?: unknown;
-    metadata?: unknown;
-  }>;
-}>;
-
-type DeckUpsertLegacyEffortMigrationOperation = Extract<
-  SyncPushOperation,
-  Readonly<{ entityType: "deck"; action: "upsert" }>
->;
-
-type StoredOutboxLegacyEffortMigrationRecord = Readonly<{
-  operation: SyncPushOperation;
-}>;
-
-function hasOwnProperty(objectValue: object, key: string): boolean {
-  return Object.prototype.hasOwnProperty.call(objectValue, key);
-}
 
 function isQuotaExceededError(error: unknown): boolean {
   return error instanceof DOMException && error.name === "QuotaExceededError";
@@ -271,476 +202,6 @@ export function readIndexedDbOpenLifecycleSnapshotForDiagnostics(): IndexedDbOpe
   return lastIndexedDbVersionChangeLifecycleSnapshot ?? lastIndexedDbOpenLifecycleSnapshot;
 }
 
-function deleteExistingStore(database: IDBDatabase, storeName: string): void {
-  if (database.objectStoreNames.contains(storeName)) {
-    database.deleteObjectStore(storeName);
-  }
-}
-
-function deleteExistingIndex(store: IDBObjectStore, indexName: string): void {
-  if (store.indexNames.contains(indexName)) {
-    store.deleteIndex(indexName);
-  }
-}
-
-function createReviewEventsIndexes(reviewEventsStore: IDBObjectStore): void {
-  if (!reviewEventsStore.indexNames.contains("workspaceId_reviewedAtClient_reviewEventId")) {
-    reviewEventsStore.createIndex(
-      "workspaceId_reviewedAtClient_reviewEventId",
-      ["workspaceId", "reviewedAtClient", "reviewEventId"],
-      { unique: false },
-    );
-  }
-}
-
-function createCardsUpdatedAtIndexes(cardsStore: IDBObjectStore): void {
-  if (!cardsStore.indexNames.contains("workspaceId_updatedAt_cardId")) {
-    cardsStore.createIndex("workspaceId_updatedAt_cardId", ["workspaceId", "updatedAt", "cardId"], { unique: false });
-  }
-}
-
-function createCardsDueAtMillisIndex(cardsStore: IDBObjectStore): void {
-  if (!cardsStore.indexNames.contains("workspaceId_dueAtMillis_cardId")) {
-    cardsStore.createIndex("workspaceId_dueAtMillis_cardId", ["workspaceId", "dueAtMillis", "cardId"], { unique: false });
-  }
-}
-
-function createCardsDueAtBucketMillisIndex(cardsStore: IDBObjectStore): void {
-  if (!cardsStore.indexNames.contains("workspaceId_dueAtBucketMillis_cardId")) {
-    cardsStore.createIndex("workspaceId_dueAtBucketMillis_cardId", ["workspaceId", "dueAtBucketMillis", "cardId"], { unique: false });
-  }
-}
-
-function createCardsFsrsLastReviewedAtMillisIndex(cardsStore: IDBObjectStore): void {
-  if (!cardsStore.indexNames.contains("workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_cardId")) {
-    cardsStore.createIndex(
-      "workspaceId_fsrsLastReviewedAtMillis_dueAtMillis_cardId",
-      ["workspaceId", "fsrsLastReviewedAtMillis", "dueAtMillis", "cardId"],
-      { unique: false },
-    );
-  }
-}
-
-function createCardsStore(database: IDBDatabase): void {
-  const cardsStore = database.createObjectStore("cards", { keyPath: ["workspaceId", "cardId"] });
-  cardsStore.createIndex("workspaceId_createdAt_cardId", ["workspaceId", "createdAt", "cardId"], { unique: false });
-  // TODO: Drop this legacy dueAt index after cards-list sorting no longer depends on the boundary string field.
-  cardsStore.createIndex("workspaceId_dueAt_cardId", ["workspaceId", "dueAt", "cardId"], { unique: false });
-  createCardsDueAtMillisIndex(cardsStore);
-  createCardsDueAtBucketMillisIndex(cardsStore);
-  createCardsUpdatedAtIndexes(cardsStore);
-  createCardsFsrsLastReviewedAtMillisIndex(cardsStore);
-}
-
-function createCardTagsStore(database: IDBDatabase): void {
-  const cardTagsStore = database.createObjectStore("cardTags", { keyPath: ["workspaceId", "cardId", "tag"] });
-  cardTagsStore.createIndex("workspaceId_tag_cardId", ["workspaceId", "tag", "cardId"], { unique: false });
-  cardTagsStore.createIndex("workspaceId_cardId_tag", ["workspaceId", "cardId", "tag"], { unique: false });
-}
-
-function createDecksStore(database: IDBDatabase): void {
-  const decksStore = database.createObjectStore("decks", { keyPath: ["workspaceId", "deckId"] });
-  decksStore.createIndex("workspaceId_createdAt_deckId", ["workspaceId", "createdAt", "deckId"], { unique: false });
-}
-
-function createMediaAssetsStore(database: IDBDatabase): void {
-  const mediaAssetsStore = database.createObjectStore("mediaAssets", { keyPath: ["workspaceId", "mediaAssetId"] });
-  mediaAssetsStore.createIndex("workspaceId_updatedAt_mediaAssetId", ["workspaceId", "updatedAt", "mediaAssetId"], { unique: false });
-}
-
-function createReviewEventsStore(database: IDBDatabase): void {
-  const reviewEventsStore = database.createObjectStore("reviewEvents", { keyPath: ["workspaceId", "reviewEventId"] });
-  createReviewEventsIndexes(reviewEventsStore);
-}
-
-function createProgressDailyCountsStore(database: IDBDatabase): void {
-  database.createObjectStore("progressDailyCounts", { keyPath: ["workspaceId", "localDate"] });
-}
-
-function createWorkspaceSettingsStore(database: IDBDatabase): void {
-  database.createObjectStore("workspaceSettings", { keyPath: "workspaceId" });
-}
-
-function createWorkspaceSyncStateStore(database: IDBDatabase): void {
-  database.createObjectStore("workspaceSyncState", { keyPath: "workspaceId" });
-}
-
-function createOutboxStore(database: IDBDatabase): void {
-  const outboxStore = database.createObjectStore("outbox", { keyPath: ["workspaceId", "operationId"] });
-  outboxStore.createIndex("workspaceId_createdAt", ["workspaceId", "createdAt"], { unique: false });
-}
-
-function createMetaStore(database: IDBDatabase): void {
-  database.createObjectStore("meta", { keyPath: "key" });
-}
-
-function upgradeToVersion4(database: IDBDatabase): void {
-  for (const storeName of [
-    "cards",
-    "cardTags",
-    "decks",
-    "progressDailyCounts",
-    "reviewEvents",
-    "workspaceSettings",
-    "workspaceSyncState",
-    "outbox",
-    "meta",
-  ]) {
-    deleteExistingStore(database, storeName);
-  }
-
-  createCardsStore(database);
-  createCardTagsStore(database);
-  createDecksStore(database);
-  createProgressDailyCountsStore(database);
-  createReviewEventsStore(database);
-  createWorkspaceSettingsStore(database);
-  createWorkspaceSyncStateStore(database);
-  createOutboxStore(database);
-  createMetaStore(database);
-}
-
-function upgradeToVersion5(database: IDBDatabase): void {
-  deleteExistingStore(database, "workspaceSyncState");
-  createWorkspaceSyncStateStore(database);
-}
-
-function upgradeToVersion6(database: IDBDatabase): void {
-  upgradeToVersion4(database);
-}
-
-function upgradeToVersion7(transaction: IDBTransaction): void {
-  const cardsStore = transaction.objectStore("cards");
-  createCardsUpdatedAtIndexes(cardsStore);
-}
-
-function upgradeToVersion8(transaction: IDBTransaction): void {
-  const reviewEventsStore = transaction.objectStore("reviewEvents");
-  createReviewEventsIndexes(reviewEventsStore);
-}
-
-function upgradeToVersion9(database: IDBDatabase): void {
-  if (database.objectStoreNames.contains("progressDailyCounts") === false) {
-    createProgressDailyCountsStore(database);
-  }
-}
-
-function normalizeStoredCardMetadataFields(record: StoredCardMetadataMigrationRecord): StoredCard {
-  return {
-    ...record,
-    cardType: normalizeCardType(record.cardType),
-    metadata: normalizeCardMetadata(record.metadata, record.createdAt),
-  };
-}
-
-function normalizeStoredCardDueAtDerivedFields(record: StoredCardDueAtMigrationRecord): StoredCard {
-  const dueAt = record.dueAt ?? null;
-  return normalizeStoredCardMetadataFields({
-    ...record,
-    dueAt,
-    dueAtMillis: deriveDueAtMillis(dueAt),
-    dueAtBucketMillis: deriveDueAtBucketMillis(dueAt),
-    fsrsLastReviewedAtMillis: normalizeStoredCardFsrsLastReviewedAtMillis(record),
-  });
-}
-
-function normalizeStoredCardLegacyEffort(record: StoredCardLegacyEffortMigrationRecord): StoredCard {
-  const { effortLevel, ...storedCardRecord } = record;
-  return {
-    ...normalizeStoredCardDueAtDerivedFields(storedCardRecord),
-    tags: appendLegacyEffortTag(record.tags, effortLevel),
-  };
-}
-
-function appendLegacyEffortLevelsToTags(
-  tags: ReadonlyArray<string>,
-  effortLevels: ReadonlyArray<LegacyEffortLevel>,
-): ReadonlyArray<string> {
-  return effortLevels.reduce(
-    (currentTags, effortLevel) => appendLegacyEffortTag(currentTags, effortLevel),
-    tags,
-  );
-}
-
-function normalizeStoredDeckLegacyEffort(record: StoredDeckLegacyEffortMigrationRecord): Deck {
-  const {
-    filterDefinition,
-    ...storedDeckRecord
-  } = record;
-
-  return {
-    ...storedDeckRecord,
-    filterDefinition: {
-      version: filterDefinition.version,
-      tags: appendLegacyEffortLevelsToTags(filterDefinition.tags, filterDefinition.effortLevels ?? []),
-    },
-  };
-}
-
-function normalizeOutboxCardUpsertLegacyEffort(
-  operation: CardUpsertLegacyEffortMigrationOperation,
-): CardUpsertOperation {
-  const { cardType, metadata, ...payloadWithoutMetadataFields } = operation.payload;
-  const metadataFields: {
-    cardType?: NonNullable<CardUpsertOperation["payload"]["cardType"]>;
-    metadata?: NonNullable<CardUpsertOperation["payload"]["metadata"]>;
-  } = {};
-  if (hasOwnProperty(operation.payload, "cardType")) {
-    metadataFields.cardType = normalizeCardType(cardType);
-  }
-  if (hasOwnProperty(operation.payload, "metadata")) {
-    metadataFields.metadata = normalizeCardMetadata(metadata, operation.payload.createdAt);
-  }
-
-  return {
-    ...operation,
-    payload: {
-      ...payloadWithoutMetadataFields,
-      ...metadataFields,
-      tags: appendLegacyEffortTag(operation.payload.tags, operation.payload.effortLevel),
-      effortLevel: "fast",
-    },
-  };
-}
-
-function normalizeOutboxDeckUpsertLegacyEffort(
-  operation: DeckUpsertLegacyEffortMigrationOperation,
-): DeckUpsertLegacyEffortMigrationOperation {
-  return {
-    ...operation,
-    payload: {
-      ...operation.payload,
-      filterDefinition: {
-        version: operation.payload.filterDefinition.version,
-        effortLevels: [],
-        tags: appendLegacyEffortLevelsToTags(
-          operation.payload.filterDefinition.tags,
-          operation.payload.filterDefinition.effortLevels,
-        ),
-      },
-    },
-  };
-}
-
-function normalizeOutboxOperationLegacyEffort(operation: SyncPushOperation): SyncPushOperation {
-  if (operation.entityType === "card" && operation.action === "upsert") {
-    return normalizeOutboxCardUpsertLegacyEffort(operation);
-  }
-
-  if (operation.entityType === "deck" && operation.action === "upsert") {
-    return normalizeOutboxDeckUpsertLegacyEffort(operation);
-  }
-
-  return operation;
-}
-
-function normalizeStoredOutboxLegacyEffort(
-  record: StoredOutboxLegacyEffortMigrationRecord,
-): StoredOutboxLegacyEffortMigrationRecord {
-  return {
-    ...record,
-    // TODO: Remove this legacy outbox shim when the backend sync wire contract drops effort.
-    operation: normalizeOutboxOperationLegacyEffort(record.operation),
-  };
-}
-
-function normalizeStoredCardFsrsLastReviewedAtMillis(
-  record: Pick<StoredCard, "fsrsLastReviewedAt"> & Readonly<{ fsrsLastReviewedAtMillis?: number | null }>,
-): number | null {
-  if (typeof record.fsrsLastReviewedAtMillis === "number" && Number.isFinite(record.fsrsLastReviewedAtMillis)) {
-    return record.fsrsLastReviewedAtMillis;
-  }
-
-  if (record.fsrsLastReviewedAt === null) {
-    return null;
-  }
-
-  return parseDueAtMillis(record.fsrsLastReviewedAt);
-}
-
-function migrateCardsDueAtDerivedFields(cardsStore: IDBObjectStore, errorPrefix: string): void {
-  const request = cardsStore.openCursor();
-  request.onerror = () => {
-    throw describeIndexedDbError(errorPrefix, request.error);
-  };
-  request.onsuccess = () => {
-    const cursor = request.result;
-    if (cursor === null) {
-      return;
-    }
-
-    cursor.update(normalizeStoredCardDueAtDerivedFields(cursor.value as StoredCardDueAtMigrationRecord));
-    cursor.continue();
-  };
-}
-
-function migrateCardsDueAtMillis(cardsStore: IDBObjectStore): void {
-  migrateCardsDueAtDerivedFields(cardsStore, "IndexedDB dueAtMillis migration failed");
-}
-
-function upgradeToVersion10(transaction: IDBTransaction): void {
-  const cardsStore = transaction.objectStore("cards");
-  createCardsDueAtMillisIndex(cardsStore);
-  migrateCardsDueAtMillis(cardsStore);
-}
-
-function migrateCardsDueAtBucketMillis(cardsStore: IDBObjectStore): void {
-  migrateCardsDueAtDerivedFields(cardsStore, "IndexedDB dueAtBucketMillis migration failed");
-}
-
-function upgradeToVersion11(transaction: IDBTransaction): void {
-  const cardsStore = transaction.objectStore("cards");
-  createCardsDueAtBucketMillisIndex(cardsStore);
-  migrateCardsDueAtBucketMillis(cardsStore);
-}
-
-function upgradeToVersion12(transaction: IDBTransaction): void {
-  const cardsStore = transaction.objectStore("cards");
-  migrateCardsDueAtDerivedFields(cardsStore, "IndexedDB dueAt sentinel migration failed");
-}
-
-function upgradeToVersion13(transaction: IDBTransaction): void {
-  const cardsStore = transaction.objectStore("cards");
-  createCardsFsrsLastReviewedAtMillisIndex(cardsStore);
-  migrateCardsDueAtDerivedFields(cardsStore, "IndexedDB fsrsLastReviewedAtMillis migration failed");
-}
-
-function upgradeToVersion14(transaction: IDBTransaction): void {
-  transaction.objectStore("progressDailyCounts").clear();
-  const metaStore = transaction.objectStore("meta");
-  const cacheStateRequest = metaStore.get(progressCacheStateKey);
-
-  cacheStateRequest.onerror = () => {
-    throw describeIndexedDbError("IndexedDB progress rating-count migration failed", cacheStateRequest.error);
-  };
-  cacheStateRequest.onsuccess = () => {
-    const cacheState = cacheStateRequest.result as ProgressCacheStateRecord | undefined;
-    if (cacheState === undefined) {
-      return;
-    }
-
-    metaStore.put({
-      ...cacheState,
-      needsRebuild: true,
-      updatedAt: new Date().toISOString(),
-    });
-  };
-}
-
-function migrateCardsLegacyEffortTags(cardsStore: IDBObjectStore, cardTagsStore: IDBObjectStore): void {
-  const request = cardsStore.openCursor();
-  request.onerror = () => {
-    throw describeIndexedDbError("IndexedDB legacy effort migration failed", request.error);
-  };
-  request.onsuccess = () => {
-    const cursor = request.result;
-    if (cursor === null) {
-      return;
-    }
-
-    const migratedCard = normalizeStoredCardLegacyEffort(cursor.value as StoredCardLegacyEffortMigrationRecord);
-    cursor.update(migratedCard);
-    if (migratedCard.deletedAt === null) {
-      for (const tag of migratedCard.tags) {
-        if (tag === "") {
-          continue;
-        }
-
-        cardTagsStore.put({
-          workspaceId: migratedCard.workspaceId,
-          cardId: migratedCard.cardId,
-          tag,
-        });
-      }
-    }
-    cursor.continue();
-  };
-}
-
-function migrateDecksLegacyEffortTags(decksStore: IDBObjectStore): void {
-  const request = decksStore.openCursor();
-  request.onerror = () => {
-    throw describeIndexedDbError("IndexedDB legacy deck effort migration failed", request.error);
-  };
-  request.onsuccess = () => {
-    const cursor = request.result;
-    if (cursor === null) {
-      return;
-    }
-
-    cursor.update(normalizeStoredDeckLegacyEffort(cursor.value as StoredDeckLegacyEffortMigrationRecord));
-    cursor.continue();
-  };
-}
-
-function migrateOutboxLegacyEffortOperations(outboxStore: IDBObjectStore): void {
-  const request = outboxStore.openCursor();
-  request.onerror = () => {
-    throw describeIndexedDbError("IndexedDB legacy outbox effort migration failed", request.error);
-  };
-  request.onsuccess = () => {
-    const cursor = request.result;
-    if (cursor === null) {
-      return;
-    }
-
-    cursor.update(normalizeStoredOutboxLegacyEffort(cursor.value as StoredOutboxLegacyEffortMigrationRecord));
-    cursor.continue();
-  };
-}
-
-function upgradeToVersion15(transaction: IDBTransaction): void {
-  const cardsStore = transaction.objectStore("cards");
-  deleteExistingIndex(cardsStore, "workspaceId_effort_createdAt_cardId");
-  deleteExistingIndex(cardsStore, "workspaceId_effort_updatedAt_cardId");
-  migrateCardsLegacyEffortTags(cardsStore, transaction.objectStore("cardTags"));
-  migrateDecksLegacyEffortTags(transaction.objectStore("decks"));
-  migrateOutboxLegacyEffortOperations(transaction.objectStore("outbox"));
-}
-
-function migrateCardsMetadataFields(cardsStore: IDBObjectStore): void {
-  const request = cardsStore.openCursor();
-  request.onerror = () => {
-    throw describeIndexedDbError("IndexedDB card metadata migration failed", request.error);
-  };
-  request.onsuccess = () => {
-    const cursor = request.result;
-    if (cursor === null) {
-      return;
-    }
-
-    cursor.update(normalizeStoredCardLegacyEffort(cursor.value as StoredCardLegacyEffortMigrationRecord));
-    cursor.continue();
-  };
-}
-
-function migrateOutboxCardMetadataOperations(outboxStore: IDBObjectStore): void {
-  const request = outboxStore.openCursor();
-  request.onerror = () => {
-    throw describeIndexedDbError("IndexedDB outbox card metadata migration failed", request.error);
-  };
-  request.onsuccess = () => {
-    const cursor = request.result;
-    if (cursor === null) {
-      return;
-    }
-
-    cursor.update(normalizeStoredOutboxLegacyEffort(cursor.value as StoredOutboxLegacyEffortMigrationRecord));
-    cursor.continue();
-  };
-}
-
-function upgradeToVersion16(transaction: IDBTransaction): void {
-  migrateCardsMetadataFields(transaction.objectStore("cards"));
-  migrateOutboxCardMetadataOperations(transaction.objectStore("outbox"));
-}
-
-function upgradeToVersion17(database: IDBDatabase): void {
-  if (database.objectStoreNames.contains("mediaAssets") === false) {
-    createMediaAssetsStore(database);
-  }
-}
-
 export function openDatabase(): Promise<IDBDatabase> {
   return new Promise((resolve, reject) => {
     const request = indexedDB.open(databaseName, databaseVersion);
@@ -755,107 +216,7 @@ export function openDatabase(): Promise<IDBDatabase> {
     request.onupgradeneeded = (event) => {
       const oldVersion = event.oldVersion;
       oldVersionDuringUpgrade = oldVersion;
-
-      if (oldVersion < 4) {
-        upgradeToVersion4(request.result);
-      }
-
-      if (oldVersion < 5) {
-        upgradeToVersion5(request.result);
-      }
-
-      if (oldVersion < 6) {
-        upgradeToVersion6(request.result);
-      }
-
-      if (oldVersion < 7) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion7(transaction);
-      }
-
-      if (oldVersion < 8) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion8(transaction);
-      }
-
-      if (oldVersion < 9) {
-        upgradeToVersion9(request.result);
-      }
-
-      if (oldVersion < 10) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion10(transaction);
-      }
-
-      if (oldVersion < 11) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion11(transaction);
-      }
-
-      if (oldVersion < 12) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion12(transaction);
-      }
-
-      if (oldVersion < 13) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion13(transaction);
-      }
-
-      if (oldVersion < 14) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion14(transaction);
-      }
-
-      if (oldVersion < 15) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion15(transaction);
-      }
-
-      if (oldVersion < 16) {
-        const transaction = request.transaction;
-        if (transaction === null) {
-          throw new Error("IndexedDB upgrade transaction is unavailable");
-        }
-
-        upgradeToVersion16(transaction);
-      }
-
-      if (oldVersion < 17) {
-        upgradeToVersion17(request.result);
-      }
+      upgradeDatabase(request.result, oldVersion, request.result.version, request.transaction);
     };
 
     request.onsuccess = () => {
